+++ IsekaiMod/Content/Classes/IsekaiProtagonist/IsekaiProtagonistClass.cs
@@ -30,7 +30,7 @@ namespace IsekaiMod.Content.Classes.IsekaiProtagonist
 
         public static void Add()
         {
-            // TODO: add better icon for Hero's presence
+            // TODO: add channel positive and channel negative energy (requires good for positive and evil for negative)
 
             // TODO: add channel negative energy to villain
             // TODO: add scaling natural armor, strength and dexterity (like animal companion) to deathsnatcher