+++ IsekaiMod/Content/Features/IsekaiProtagonist/CharacterDevelopment/CharacterDevelopmentSelection.cs
@@ -3,18 +3,16 @@ using IsekaiMod.Utilities;
 using Kingmaker.Blueprints;
 using Kingmaker.Blueprints.Classes;
 using Kingmaker.Blueprints.Classes.Selection;
+using UnityEngine;
 
 namespace IsekaiMod.Content.Features.IsekaiProtagonist.CharacterDevelopment
 {
     class CharacterDevelopmentSelection
     {
+        private static readonly Sprite Icon_Discovery = Resources.GetBlueprint<BlueprintFeatureSelection>("cd86c437488386f438dcc9ae727ea2a6").m_Icon;
         public static void Add()
         {
-            var IsekaiProtagonistBonusFeatSelection = Resources.GetModBlueprint<BlueprintFeatureSelection>("IsekaiProtagonistBonusFeatSelection");
             var IsekaiProtagonistTalentSelection = Resources.GetModBlueprint<BlueprintFeatureSelection>("IsekaiProtagonistTalentSelection");
-
-            // Feature
-            var Icon_Discovery = Resources.GetBlueprint<BlueprintFeatureSelection>("cd86c437488386f438dcc9ae727ea2a6").m_Icon;
             var CharacterDevelopmentSelection = Helpers.CreateBlueprint<BlueprintFeatureSelection>("CharacterDevelopmentSelection", bp => {
                 bp.SetName("Character Development");
                 bp.SetDescription("As you increase your level, you can select powerful character development feats.");
@@ -23,14 +21,8 @@ namespace IsekaiMod.Content.Features.IsekaiProtagonist.CharacterDevelopment
                 bp.IsClassFeature = true;
 
                 // The rest of the character development feats are added in later
-                bp.m_Features = new BlueprintFeatureReference[] {
-                    IsekaiProtagonistBonusFeatSelection.ToReference<BlueprintFeatureReference>(),
-                    IsekaiProtagonistTalentSelection.ToReference<BlueprintFeatureReference>()
-                };
-                bp.m_AllFeatures = new BlueprintFeatureReference[] {
-                    IsekaiProtagonistBonusFeatSelection.ToReference<BlueprintFeatureReference>(),
-                    IsekaiProtagonistTalentSelection.ToReference<BlueprintFeatureReference>()
-                };
+                bp.m_Features = new BlueprintFeatureReference[] { IsekaiProtagonistTalentSelection.ToReference<BlueprintFeatureReference>() };
+                bp.m_AllFeatures = new BlueprintFeatureReference[] { IsekaiProtagonistTalentSelection.ToReference<BlueprintFeatureReference>() };
             });
         }
         public static void AddToSelection(BlueprintFeature feature)